requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
python-magic==0.4.27
email-validator==2.1.0.post1
gunicorn==21.2.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0 